+++ src/main/java/janggi/JanggiGame.java
@@ -27,12 +27,9 @@ public class JanggiGame {
 
                 Position selectedPiecePosition = inputView.selectPiece();
 
-                List<Position> reachablePositions = board.computeReachableDestination(selectedPiecePosition);
-                outputView.printReachableDestinations(reachablePositions);
+                List<Position> reachablePositions = computeReachableDestinations(selectedPiecePosition);
 
-                Position destination = inputView.askMovableDestination();
-                Piece catchedPiece = board.moveOrCatchPiece(selectedPiecePosition, destination, reachablePositions);
-                outputView.printMoveResult(catchedPiece);
+                Piece catchedPiece = processMove(selectedPiecePosition, reachablePositions);
 
                 if (board.checkGameIsOver(catchedPiece)) {
                     break;
@@ -42,4 +39,18 @@ public class JanggiGame {
             }
         }
     }
+
+    private List<Position> computeReachableDestinations(final Position selectedPiecePosition) {
+        List<Position> reachablePositions = board.computeReachableDestination(selectedPiecePosition);
+        outputView.printReachableDestinations(reachablePositions);
+        return reachablePositions;
+    }
+
+    private Piece processMove(final Position selectedPiecePosition, final List<Position> reachablePositions) {
+        Position destination = inputView.askMovableDestination();
+        Piece catchedPiece = board.moveOrCatchPiece(selectedPiecePosition, destination, reachablePositions);
+        outputView.printMoveResult(catchedPiece);
+        return catchedPiece;
+    }
+
 }
