+++ src/test/java/janggi/piece/SoldierTest.java
@@ -1,6 +1,5 @@
 package janggi.piece;
 
-import janggi.Side;
 import janggi.board.Position;
 import janggi.board.Route;
 import org.junit.jupiter.api.DisplayName;
